cleaned up save_app.py file
